test-automation-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.133.0"}
python-dotenv = "^1.2.1"
google-genai = "^1.64.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
